User Authentication Implemented / added routes folder
I personally like to put routes in another file, one for each controller. For me it makes debugging easier and the code more readable but I'm happy to implement this whichever way we agree upon. I left the other routes and used an example of this structure with the User routes.